% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\docType{data}
\name{algorithms}
\alias{algorithms}
\alias{working_model_G_one}
\alias{working_model_G_two}
\alias{working_model_G_three}
\alias{working_model_Q_one}
\alias{kknn_algo}
\alias{kknn_grid}
\alias{kknn_control}
\title{Algorithms for the estimation of Gbar and Qbar}
\format{Either  a  \code{list}   with  entries  'model',  'formula'  and
  'type_of_preds'  (with \code{attribue}  'ML'  set to  \code{FALSE}) or  a
  \code{list} with entries 'algo' and 'type_of_preds' (with \code{attribue}
  'ML' set to \code{TRUE}).}
\usage{
data(guided.tour.tmle)

working_model_G_one

working_model_G_two

working_model_G_three

working_model_Q_one

kknn_algo

kknn_grid

kknn_control
}
\description{
Algorithms  for the  estimation of  'Gbar' and  'Qbar'. Some  are based  on
parametric  working  models  and  the  \code{\link{glm}}  procedure  (their
\code{attribute}  'ML' is  set to  \code{FALSE}). Another  is based  on the
k-nearest  neighbors  non-parametric  procedure as  implement  in  'method'
\code{kknn} of  the 'caret'  package (its \code{attribute}  'ML' is  set to
\code{TRUE}).
}
\examples{

(working_model_G_one)
(kknn_algo)

}
\references{
Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).
}
\keyword{datasets}
