% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{alter}
\alias{alter}
\alias{alter.LAW}
\title{Alters an object of class LAW}
\arguments{
\item{this}{An object of class \code{LAW}.}

\item{\dots}{...}
}
\value{
something
}
\description{
Alters an object of class \code{LAW} by ...
}
\seealso{
\code{\link{sample_from}},       \code{\link{reveal}},
  \code{\link{evaluate_psi}}, \code{\link{evaluate_eic}}
}
