% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R, R/misc.R
\docType{data}
\name{convenient_functions}
\alias{convenient_functions}
\alias{compute_lGbar_hatAW}
\alias{compute_Qbar_hatAW}
\alias{wrapper}
\alias{label}
\title{Convenient functions.}
\format{A \code{function}.}
\usage{
data(guided.tour.tmle)

compute_lGbar_hatAW(A, W, Gbar_hat, threshold = 0.05)

compute_Qbar_hatAW(Y, A, W, Qbar_hat, blip = FALSE)

wrapper(fit)

label(xx, sample_size = c(1000, 2000))
}
\arguments{
\item{xx}{A \code{vector}.}

\item{sample_size}{A \code{vector} of \code{integer}s.}
}
\value{
A \code{vector} of \code{characters}.
}
\description{
Convenient functions.
}
\details{
Let 'L'  be the length of 'sample_size' and  'Q'be the quotient in
  the Euclidean division of the  length of 'xx' by 'sum(sample_size)'.  The
  \code{function} \code{label}  creates 'Q  x L'  different strings  of the
  form 'q_l'.   Each 'q_l' string  is repeated 'sample_size[l]'  times. The
  function is used to make easily sub-data sets of a data set.
}
\references{
Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).

Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).

Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).
}
