% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_Gbar}
\alias{estimate_Gbar}
\title{Does this}
\usage{
estimate_Gbar(dat, algorithm, ...)
}
\arguments{
\item{dat}{...}

\item{algorithm}{...}

\item{...}{Additional parameters passed to the algorithm.}
}
\value{
something
}
\description{
Does this.
}
\examples{

## do this and that
}
\seealso{
\code{\link{estimate_QW}}, \code{\link{estimate_Qbar}}
}
