% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_Qbar}
\alias{estimate_Qbar}
\title{Does this}
\usage{
estimate_Qbar(dat, algorithm, ...)
}
\arguments{
\item{dat}{...}

\item{algorithm}{...}

\item{...}{Additional parameters passed to the algorithm.}
}
\value{
something
}
\description{
Does this.
}
\examples{

## do this and that (refer to working models and knn)

}
\seealso{
\code{\link{estimate_QW}}, \code{\link{estimate_Gbar}}
}
