% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_QW}
\alias{estimate_QW}
\title{Estimates the marginal law of W non-parametrically}
\usage{
estimate_QW(dat)
}
\arguments{
\item{dat}{...}
}
\value{
A  \code{tibble}  with  columns named  'value'  and 'weight'  that
  describes the empirical law of 'W' in data set 'dat'.
}
\description{
Estimates the marginal law of W non-parametrically.
}
\examples{

## do this and that
}
\seealso{
\code{\link{estimate_Gbar}}, \code{\link{estimate_Qbar}}
}
