% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{evaluate_eic}
\alias{evaluate_eic}
\alias{evaluate_eic.LAW}
\title{Evaluates the efficient influence curve of Psi at an object of class LAW}
\arguments{
\item{this}{An object of class \code{LAW}.}

\item{\dots}{Additional parameters possibly needed to  fully characterize the
law.}
}
\value{
A \code{function},  the efficient  infuence curve  of statistical
  mapping \eqn{\Psi} evaluated at the law described by the \code{LAW} object
  \code{this}.
}
\description{
Evaluates  at   efficient  influence  curve  of   the  statistical  mapping
\eqn{\Psi} at the law for (W,A,Y) in [0,1] x {0,1} x [0,1] characterized by
an object  of class \code{LAW}.   The mapping \eqn{\Psi} and  its efficient
influence  curve  at  a  law  \eqn{P} are  given  by  \deqn{\Psi(P)  =  E_P
{Qbar(1,W) -  Qbar(0,W)}} with \eqn{Qbar(A,W)  = E_P (Y|A,W)}  and \eqn{D^*
(P)  = D_1^*  (P) +  D_2^*(P)}  where \deqn{D_1^*(P)(W,A,Y)  = Qbar(1,W)  -
Qbar(0,W) - \Psi(P)} and  \deqn{D_2^*(P)(W,A,Y) = \frac{2A-1}{Gbar(A,W)} (Y
- Qbar(A,W)).}
}
\details{
The evaluation can be performed if one  has built the object or if one acts
as an  \emph{oracle}.  Works only if  all the relevant features  of the law
are characterized.
}
\seealso{
\code{\link{sample_from}},       \code{\link{reveal}},
  \code{\link{alter}}, \code{\link{evaluate_psi}}
}
