% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{reveal}
\alias{reveal}
\alias{reveal.LAW}
\alias{@seealso}
\alias{\code{\link{sample_from}},}
\alias{\code{\link{update}},}
\alias{\code{\link{evaluate_psi}},}
\alias{\code{\link{evaluate_eic}}}
\title{Reveals an object of class LAW}
\arguments{
\item{this}{An object of class \code{LAW}}

\item{\dots}{Additional parameters possibly needed to  fully determine the
law.}
}
\value{
A \code{list} with tags\itemize{\item  'QW', marginal law of 'W', a
  \code{function}  (the  density) or  a  \code{tibble}  with columns  named
  'value'   and  'weight'   (a  discrete   law)\item  'Gbar',   conditional
  probability  that  'A=1'  given  'W',  a  \code{function}  \item  'Qbar',
  conditional  mean of  'Y' given  '(A,W)', a  \code{function} \item  'qY',
  conditional density  of 'Y' given  '(A,W)', a \code{function}.}   Each of
  them equals 'NA' if it is not determined.
}
\description{
Reveals an object of class \code{LAW} by exhibiting a selection of relevant features. Can be performed if one has built the object or if one acts as an oracle.
}
