% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{sample_from}
\alias{sample_from}
\alias{sample_from.LAW}
\title{Samples from an object of class LAW.}
\arguments{
\item{this}{An object of class \code{LAW}}

\item{n}{An \code{integer}, the wished sample size. Defaults to 'n = 1'.}

\item{\dots}{Additional parameters possibly needed to fully determine the law.}
}
\value{
Either \code{NULL} if the law is undetermined or the data set sampled, a \code{tibble} or \code{data.frame} with columns named 'W', 'A' and 'Y'.
}
\description{
Samples from an object of class \code{LAW} when it is not underdetermined.
}
\seealso{
\code{\link{reveal}}, \code{\link{update}}, \code{\link{evaluate_psi}}, \code{\link{evaluate_eic}}
}
