% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{trim_glm_fit}
\alias{trim_glm_fit}
\title{Trims the fat from a 'glm' object}
\usage{
trim_glm_fit(x)
}
\arguments{
\item{x}{A \code{glm} fit.}
}
\value{
Object 'fit' stripped of components that are not required to make predictions.
}
\description{
Imported from the 'glm' 'model' of the 'caret' package, trims the fat from a 'glm' object.
}
\examples{

## Dobson (1990) Page 93: Randomized Controlled Trial (see 'example(glm)'):
counts <- c(18,17,15,20,10,20,25,13,12)
outcome <- gl(3,1,9)
treatment <- gl(3,3)
dat <- data.frame(treatment, outcome, counts)
fit <- glm(counts ~ outcome + treatment, family = poisson())
fit_light <- trim_glm_fit(fit)
if (requireNamespace("pryr", quietly = TRUE)) {
  pryr::object_size(fit)
  pryr::object_size(fit_light)
} else {
  object.size(fit)
  object.size(fit_light)
}
predict(fit_light, newdata = dat)
}
