% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{as.character}
\alias{as.character}
\alias{as.character.LAW}
\title{Prints an object of class LAW}
\arguments{
\item{x}{An object of class \code{LAW}.}

\item{\dots}{Not used.}
}
\value{
A character string recalling what can be done with an object of class \code{LAW}.
}
\description{
Prints an object of class LAW.
}
