% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R, R/misc.R
\docType{data}
\name{convenient_functions}
\alias{convenient_functions}
\alias{compute_lGbar_hatAW}
\alias{compute_Qbar_hatAW}
\alias{wrapper}
\alias{label}
\title{Convenient functions.}
\format{A \code{function}.}
\usage{
data(guided.tour.tmle)

compute_lGbar_hatAW(A, W, Gbar_hat, threshold = 0.05)

compute_Qbar_hatAW(Y, A, W, Qbar_hat, blip = FALSE)

wrapper(fit)

label(xx, sample_size = c(1000, 2000))
}
\description{
Convenient functions.
}
\examples{

label(1:17, c(2, 3))

}
\references{
Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).

Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).
}
