% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimate_Qbar}
\alias{estimate_Qbar}
\title{Estimates the Qbar feature of an object of class LAW}
\usage{
estimate_Qbar(dat, algorithm, ...)
}
\arguments{
\item{dat}{The learning data  set. Must have the  same form as a  data set
produced    by    an    object   of    \code{class}    \code{LAW}    (see
'?guided.tour.tmle').}

\item{algorithm}{The  algorithm.  See the  built-in algorithms  by running
'data(algorithms)'.}

\item{...}{Additional parameters passed to the algorithm.}
}
\value{
The  result of  the fit  output by  the algorithm  trained on  the
  learning data set.
}
\description{
Given a data set  consisting of realizations of (W,A,Y) in  [0,1] x {0,1} x
[0,1]  drawn   from  a  common  law,   \code{estimate_Qbar}  estimates  the
conditional expectation of  Y given (A,W), the so called  'Qbar' feature of
the law.
}
\examples{

## create an experiment and draw a data set from it
invisible(example(guided.tour.tmle))
obs <- sample_from(experiment, n = 250)

## load the built-in algorithms
data(algorithms)
working_model_Q_one

## estimate 'Qbar'
Qbar_hat <- estimate_Qbar(obs, working_model_Q_one)

}
\references{
Benkeser & Chambaz, "A Guided Tour in Targeted Learning Territory" (2018).
}
\seealso{
\code{\link{estimate_QW}} to estimate the marginal distribution of
  W,  \code{\link{estimate_Gbar}} to  estimate the  conditional probability
  that A=1 given W.
}
