% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{evaluate_psi}
\alias{evaluate_psi}
\alias{evaluate_psi.LAW}
\title{Evaluates the statistical mapping Psi at an object of class LAW}
\arguments{
\item{this}{An object of class \code{LAW}.}

\item{\dots}{Additional parameters possibly needed to  fully characterize the
law.}
}
\value{
A \code{function},  the efficient  infuence curve  of statistical
  mapping \eqn{\Psi} evaluated at the law described by the \code{LAW} object
  \code{this}.
}
\description{
Evaluates at the law for (W,A,Y) in  [0,1] x {0,1} x [0,1] characterized by
an object of  class \code{LAW} the statistical mapping  \eqn{\Psi} given by
\deqn{\Psi(P) = E_P {Qbar(1,W)  - Qbar(0,W)}} where \eqn{Qbar(A,W)
= E_P (Y|A,W)}. Can be performed if one has built the object or if one acts
as an  \emph{oracle}.  Works only if  all the relevant features  of the law
are characterized.
}
\seealso{
\code{\link{sample_from}},       \code{\link{reveal}},
  \code{\link{alter}}, \code{\link{evaluate_eic}}
}
