% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{LAW}
\alias{LAW}
\title{Initializes an object of class LAW}
\description{
Initializes an object of \code{class} \code{LAW}.
}
\seealso{
Other methods and functions for LAW objects: \code{\link{alter}},
  \code{\link{as.character}}, \code{\link{copy}},
  \code{\link{evaluate_eic}}, \code{\link{evaluate_psi}},
  \code{\link{evaluate_remainder}},
  \code{\link{get_feature}}, \code{\link{reveal}},
  \code{\link{sample_from}}
}
\concept{methods and functions for LAW objects}
