% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAW.R
\name{sample_from}
\alias{sample_from}
\alias{sample_from.LAW}
\alias{@seealso}
\alias{\code{\link{reveal}},}
\alias{\code{\link{alter}},}
\alias{\code{\link{evaluate_psi}},}
\alias{\code{\link{evaluate_eic}}}
\title{Samples from an object of class LAW}
\arguments{
\item{this}{An object of class \code{LAW}.}

\item{n}{An \code{integer}, the wished sample size. Defaults to 'n = 1'.}

\item{\dots}{Additional parameters  possibly needed to  fully characterize
the law.}
}
\value{
Either  \code{NULL} if  the law  is undetermined  or the  data set
  sampled, a \code{tibble} or \code{data.frame} with columns named 'W', 'A'
  and 'Y'.
}
\description{
Samples data of the form (W,A,Y) in [0,1] x {0,1} x [0,1] from an object of
class \code{LAW} when it fully characterizes a law.
}
